+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Calendar Versioning](https://calver.org/) of
 the following form: YYYY.0M.0D.
 
+## [2025.10.09]
+
+### Added
+
+- Designed a proof of concept for the StageLight component
+- Designed three kernel methods, and one secondary method
+- Designed a main method to prove concept works
+
 ## [2025.09.18]
 
 ### Added